Updated x/time axis labeling
Not sure if it's xarray (0.20+) or cftime (1.5.1+) where this issue originated, but how time information in xarray metadata appears to have changed following xarray 0.19, causing matplotlib to not like how the dates were formatted and thus to provide an OverflowError: int too big to convert error when trying to label time axes. Changing from variable.Time to variable.Time.values (converting from xarray metadata to numpy array) fixes this issue for the newest versions of xarray/cftime/matplotlib and is backward-compatible.

More info on a related instance of this error message: https://stackoverflow.com/questions/66659928/overflowerror-int-too-big-to-convert-when-formatting-date-on-pandas-series-plot.